refactor: split oversized book_detail files under the 800-line cap
`physical.rs` (855) and `book_detail/mod.rs` (836) both crossed the
~800-line file-shape soft cap when the physical collection & wishlist UI
landed. Pure relocation — no markup, selector, or behaviour change.

- `physical.rs` 855 -> 643: its inline `mod tests` and `mod render_tests`
  move to a sibling `physical/tests.rs`, per the rule that non-trivial
  unit tests live in `<mod>/tests.rs`. `render_tests` becomes a nested
  module gated on `feature = "server"` alone, since the parent is already
  `#[cfg(test)]`.
- `mod.rs` 836 -> 561: the loaded-book view composition
  (`kicker_label`, `series_label`, `build_crumbs`, `LoadedBookView`,
  `derive_loaded_view`, `RailButtons`, `LoadedCtx`, `render_loaded`)
  moves to a sibling `view.rs` with its label tests in `view/tests.rs`.
  `mod.rs` keeps the data fetch, signal groups, and the shared markup
  primitives. `derive_loaded_view` / `LoadedBookView` are re-exported
  only under `feature = "mobile"`, the sole external consumer.
