+++ frontend/src/pages/book_detail/mod.rs
@@ -1,7 +1,8 @@
 //! Book detail page — Atrium "Cinematic" shell.
 //!
-//! Owns the data fetch and shared markup primitives; section bodies live in
-//! [`hero`], [`body`], and [`merge`].
+//! Owns the data fetch and shared markup primitives; the loaded-book view
+//! composition lives in [`view`], and section bodies live in [`hero`],
+//! [`body`], and [`merge`].
 
 use dioxus::prelude::*;
 use dioxus_router::Link;
@@ -19,6 +20,7 @@ mod journal_editor;
 mod merge;
 mod rating;
 mod read_status;
+mod view;
 
 // Web renders the split hero + body-grid; mobile re-flows the same loaded-book
 // data into a single-column surface. The web-only sections aren't compiled on
@@ -36,10 +38,12 @@ mod offline;
 #[cfg(not(feature = "mobile"))]
 mod physical;
 
-#[cfg(not(feature = "mobile"))]
-use body::{BdAuthorCluster, BdBodyMain, BdPageCtx, BdRailSection};
-#[cfg(not(feature = "mobile"))]
-use hero::BdHeroSection;
+use view::{render_loaded, LoadedCtx, RailButtons};
+// Only `mobile.rs` (compiled solely under this feature) re-derives the loaded
+// view itself; the web branch calls `derive_loaded_view` directly inside
+// `view::render_loaded`, so this re-export would otherwise be unused there.
+#[cfg(feature = "mobile")]
+use view::{derive_loaded_view, LoadedBookView};
 
 /// Book detail page shell: fetches metadata then hands off to `render_loaded`.
 #[component]
@@ -219,17 +223,6 @@ struct MergeSignals {
     result: Signal<Option<MergeBooksResult>>,
     undo_error: Signal<Option<String>>,
     refresh: Signal<u32>,
-}
-
-/// Prebuilt web-only rail action buttons (Merge / Delete), threaded from
-/// [`render_book_shell`] into [`render_loaded`]. Grouped only to keep that
-/// call site under clippy's argument cap; both are always `None` on mobile,
-/// and mobile discards the whole struct rather than reading its fields (see
-/// [`LoadedBookView`] for the same pattern).
-#[cfg_attr(feature = "mobile", allow(dead_code))]
-struct RailButtons {
-    merge: Option<Element>,
-    delete: Option<Element>,
 }
 
 /// Assemble the loaded-book view: derive the admin flag, build the (web-only)
@@ -473,244 +466,6 @@ fn poll_suggestions_until_resolved(
     });
 }
 
-fn kicker_label(year: &str) -> String {
-    if year.is_empty() {
-        "Book".to_string()
-    } else {
-        format!("Book · {year}")
-    }
-}
-
-fn series_label(series: Option<&str>, index: Option<&str>) -> Option<String> {
-    match (series, index) {
-        (Some(s), Some(i)) => Some(format!("{s} #{i}")),
-        (Some(s), None) => Some(s.to_string()),
-        _ => None,
-    }
-}
-
-/// Build the breadcrumb item list for a loaded book.
-fn build_crumbs(
-    b: &EbookMetadata,
-    title: &str,
-    primary_author: &str,
-    series_label: &Option<String>,
-) -> Vec<BdCrumbItem> {
-    let mut crumbs = vec![BdCrumbItem {
-        text: "Home".to_string(),
-        target: Some(Route::Landing {}),
-    }];
-    if !primary_author.is_empty() {
-        let author_route = b
-            .creators
-            .first()
-            .and_then(|c| c.id)
-            .map(|id| Route::AuthorDetail { id });
-        crumbs.push(BdCrumbItem {
-            text: primary_author.to_string(),
-            target: author_route,
-        });
-    }
-    if let Some(label) = series_label.clone() {
-        let series_route = b.series_id.map(|id| Route::SeriesDetail { id });
-        crumbs.push(BdCrumbItem {
-            text: label,
-            target: series_route,
-        });
-    }
-    crumbs.push(BdCrumbItem {
-        text: title.to_string(),
-        target: None,
-    });
-    crumbs
-}
-
-/// Pre-derived strings + flags ready to feed the loaded-book sections.
-/// Split out of [`render_loaded`] so the rsx body stays a thin composition
-/// of named sub-components. Mobile reads a subset (no breadcrumb / author-id
-/// cluster), so some fields are unused there.
-#[cfg_attr(feature = "mobile", allow(dead_code))]
-struct LoadedBookView {
-    title: String,
-    primary_author: String,
-    author_id: Option<i64>,
-    authors_line: String,
-    kicker: String,
-    series: Option<String>,
-    accent_style: String,
-    has_audio: bool,
-    has_ebook: bool,
-    crumbs: Vec<BdCrumbItem>,
-}
-
-/// Compute the per-section display fields from the loaded book.
-fn derive_loaded_view(b: &EbookMetadata) -> LoadedBookView {
-    let title = b.display_title();
-    let primary_author = b
-        .creators
-        .first()
-        .map(|c| c.name.clone())
-        .unwrap_or_default();
-    let author_id = b.creators.first().and_then(|c| c.id);
-    let authors_line = b
-        .creators
-        .iter()
-        .map(|c| c.name.clone())
-        .collect::<Vec<_>>()
-        .join(", ");
-    let year = b
-        .published
-        .as_deref()
-        .and_then(|p| p.get(0..4))
-        .unwrap_or("")
-        .to_string();
-    let kicker = kicker_label(&year);
-    let series = series_label(b.series.as_deref(), b.series_index.as_deref());
-    let accent_style = b
-        .accent
-        .as_deref()
-        .map(|a| format!("--accent: {a};"))
-        .unwrap_or_default();
-    let has_audio = b
-        .formats
-        .iter()
-        .any(|f| f.eq_ignore_ascii_case("m4b") || f.eq_ignore_ascii_case("mp3"));
-    let has_ebook = b
-        .formats
-        .iter()
-        .any(|f| f.eq_ignore_ascii_case("epub") || f.eq_ignore_ascii_case("pdf"));
-    let crumbs = build_crumbs(b, &title, &primary_author, &series);
-    LoadedBookView {
-        title,
-        primary_author,
-        author_id,
-        authors_line,
-        kicker,
-        series,
-        accent_style,
-        has_audio,
-        has_ebook,
-        crumbs,
-    }
-}
-
-/// Server URL + admin flag + book-refetch signal threaded into
-/// [`render_loaded`]. Grouped so the call site stays under clippy's argument
-/// cap; `Copy`-free (`server_url` is a `String`), so it's passed by value once.
-struct LoadedCtx {
-    server_url: String,
-    is_admin: bool,
-    refresh: Signal<u32>,
-}
-
-/// Render the fully-loaded book detail view.
-fn render_loaded(
-    b: EbookMetadata,
-    description: DescriptionSignals,
-    author_books: Vec<EbookMetadata>,
-    rail: RailButtons,
-    suggestions: Option<SuggestionsResponse>,
-    ctx: LoadedCtx,
-) -> Element {
-    let LoadedCtx {
-        server_url,
-        is_admin,
-        refresh,
-    } = ctx;
-    // Mobile re-flows the same loaded data into a single column; the web body
-    // (hero + rail + suggestions + merge UI) isn't rendered there.
-    #[cfg(feature = "mobile")]
-    let out = {
-        // The merge and delete affordances stay web-only; the discovery blocks
-        // (author cluster + suggestions) now render on mobile too. The physical
-        // panel is web-only too (issue #1186 scope), so `refresh` is unused here.
-        let _ = rail;
-        let _ = refresh;
-        mobile::render_loaded_mobile(mobile::MobileBookView {
-            b,
-            author_books,
-            suggestions,
-            is_admin,
-            server_url,
-            description,
-        })
-    };
-
-    #[cfg(not(feature = "mobile"))]
-    let out = {
-        // Web keeps its own local copy inside `BdTitleCol` (hero.rs), a real
-        // `#[component]` that gets a fresh scope per mount — no hook-order
-        // risk there, so this param is unused on this target.
-        let _ = description;
-        let RailButtons {
-            merge: merge_button,
-            delete: delete_button,
-        } = rail;
-        let LoadedBookView {
-            title,
-            primary_author,
-            author_id,
-            authors_line,
-            kicker,
-            series,
-            accent_style,
-            has_audio,
-            has_ebook,
-            crumbs,
-        } = derive_loaded_view(&b);
-
-        let uuid = b.unique_identifier.clone().unwrap_or_default();
-        // Extract the physical-panel inputs before `b` is moved into the rail.
-        let is_fileless = b.formats.is_empty();
-        let isbn13 = b.isbn13.clone();
-        let panel_author = primary_author.clone();
-        rsx! {
-            div { class: "bd-root", style: "{accent_style}",
-                BdHeroSection {
-                    b: b.clone(),
-                    title: title.clone(),
-                    kicker,
-                    crumbs,
-                    avail: hero::Availability {
-                        has_ebook,
-                        has_audio,
-                    },
-                }
-                physical::BdPhysicalPanel {
-                    uuid: uuid.clone(),
-                    is_fileless,
-                    isbn: isbn13,
-                    title: title.clone(),
-                    author: panel_author,
-                    refresh,
-                }
-                section { class: "bd-body-grid",
-                    BdBodyMain {
-                        uuid: uuid.clone(),
-                        title: title.clone(),
-                        author: BdAuthorCluster { primary_author, author_id, author_books },
-                        suggestions,
-                        ctx: BdPageCtx { server_url, is_admin },
-                    }
-                    BdRailSection {
-                        b,
-                        title,
-                        authors_line,
-                        series,
-                        merge_button,
-                        delete_button,
-                    }
-                }
-                div { class: "bd-footer",
-                    Link { to: Route::Landing {}, class: "btn", "Back to library" }
-                }
-            }
-        }
-    };
-
-    out
-}
-
 // Page-local primitives. None of these introduce business logic — they're
 // markup-only adapters so the page reads as a composition of named blocks
 // rather than nested rsx.
@@ -802,35 +557,5 @@ pub(super) fn BdInsightCell(label: String, value: String) -> Element {
             div { class: "mono bd-insight-label", "{label}" }
             div { class: "bd-insight-value", "{value}" }
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn kicker_label_renders_year_when_present() {
-        assert_eq!(kicker_label("2021"), "Book · 2021");
-    }
-    #[test]
-    fn kicker_label_falls_back_to_book_when_year_empty() {
-        assert_eq!(kicker_label(""), "Book");
-    }
-    #[test]
-    fn series_label_formats_name_and_index() {
-        assert_eq!(
-            series_label(Some("Dune"), Some("2")),
-            Some("Dune #2".into())
-        );
-    }
-    #[test]
-    fn series_label_without_index_is_just_name() {
-        assert_eq!(series_label(Some("Dune"), None), Some("Dune".into()));
-    }
-    #[test]
-    fn series_label_absent_series_is_none() {
-        assert_eq!(series_label(None, Some("2")), None);
-        assert_eq!(series_label(None, None), None);
     }
 }
